Make (0, 0) the center of the image, and position elements by center.
This makes it a easier to place elements; we can now use center()
and don't have to worry where the top left corner is going to be
in relation to the top left corner of the field we're placing
the element.

For text, we still use x and y to position, but with proper use
of CSS, it's the center we're placing (though we still want to
shift it half a pixel).