Moving towards more flexibility of which part of the infinite board to show.
We're no longer assuming we're going to cover the value 1 .. N^2, for
some odd N.

Hence, max_x doesn't have to be -min_x, nor max_x be - min_y.

Furthermore, to know whether a value is outside of the box, we cannot
just check whether it exceeds the max_value.