Clean up CSS.
We don't need to give divs two classes, one for the common styles,
and one for the things which are different. We can just list both
of them in the style sheets for things they have in common, and make
seperate sections for the ones that aren't shared.

Added some comments as well.