Slide for the (Xiangqi) horse.
This requires some additional hacking. The horse makes a knight like
jump, but it passes through square which cannot be occupied.
So, we made an adjustment to the movement animation, we can now just
a path which is a combination of shorter line segments.

For each pass through square, we drop a small dark green circle.

For now, we assume (Javascript side) there can only be one pass through square.